Código corrigido
