+++ AlunoCreate.java
@@ -1,0 +1,13 @@
+package secretaria;
+
+import java.util.ArrayList;
+
+import javax.swing.JOptionPane;
+
+public class AlunoCreate {
+
+	// Create
+	public ArrayList<Aluno> listaAlunos = new ArrayList<>();
+	
+
+}
