--- Escola/src/secretaria/DeleteAluno.java
@@ -1,5 +1,0 @@
-package secretaria;
-
-public class DeleteAluno {
-
-}
