--- Escola/src/secretaria/TurmaSegundoA.java
@@ -1,8 +1,0 @@
-package secretaria;
-
-public class TurmaSegundoA {
-
-	//create
-	
-	
-}
