--- Escola/src/secretaria/UpdateAluno.java
@@ -1,5 +1,0 @@
-package secretaria;
-
-public class UpdateAluno {
-
-}
