+++ RegistroDeTurmas.java
@@ -4,11 +4,12 @@ import java.util.ArrayList;
 
 public class RegistroDeTurmas {
 	
-	ArrayList<String> listaAlunos = new ArrayList<>();
+	ArrayList<Aluno> listaAlunos = new ArrayList<>();
 	
 	public static void main(String[] args) {
 		AlunoCreate aluno = new AlunoCreate();
-		aluno.criarAluno();
+		
+	
 		
 
 	}
